+++ poster-recommend/src/main/java/com/trilemon/boss/poster/template/dao/PosterTemplateFestivalDAO.java
@@ -3,13 +3,13 @@ package com.trilemon.boss.poster.template.dao;
 import com.trilemon.boss.poster.template.model.PosterTemplateFestival;
 
 public interface PosterTemplateFestivalDAO {
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     void insert(PosterTemplateFestival record);
 
     void insertSelective(PosterTemplateFestival record);
 
-    PosterTemplateFestival selectByPrimaryKey(Integer id);
+    PosterTemplateFestival selectByPrimaryKey(Long id);
 
     int updateByPrimaryKeySelective(PosterTemplateFestival record);
